adjusting the seed with our logic

+++ db/migrate/20200601061016_create_offer_rates.rb
@@ -1,7 +1,7 @@
 class CreateOfferRates < ActiveRecord::Migration[6.0]
   def change
     create_table :offer_rates do |t|
-      t.decimal :rate
+      t.integer :rate
       t.references :user, null: false, foreign_key: true
       t.references :offer, null: false, foreign_key: true
